requirements.txt
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.51.0
django-bootstrap4==22.3
django-crispy-forms==1.14.0
django_debug_toolbar==3.8.1
gunicorn==20.1.0
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
whitenoise==5.3.0
